refactor complete
made code cleaner, reusable, easier to read and maintain
