+++ public/js/modules/helpers.js
@@ -1,4 +1,5 @@
 import dompurify from 'dompurify';
+
 const corsProxy = 'https://stock-cors.herokuapp.com/';
 
 
@@ -23,6 +24,16 @@ function generateURL(name) {
 }
 
 
+function generateHTML(name) {
+  return `
+  <div class="stock-details">
+    <h2>${name}</h2>
+    <button class="danger" id="${name}">Remove</button>
+  </div>
+`;
+}
+
+
 function formatChartData(result) {
   return result.dataset.data.map((element) => {
     // used const because variable assignment does not change; only the value does
@@ -38,4 +49,4 @@ function sanitizeData(data) {
 }
 
 
-export { generateURL, formatChartData, sanitizeData };
+export { generateURL, generateHTML, formatChartData, sanitizeData };
