+++ views/index.pug
@@ -23,7 +23,7 @@ html(lang="en")
       form(action="" class="add-stock-form")
         p Sync across clients in realtime
         input(id="stock-name" name="company" placeholder="enter stock code")
-        input(type="submit" class="success" value="add")
+        input(type="submit" class="success" value="Add")
 
 
     script var stockOptions = []
